+++ drivers/pci/xen-pcifront.c
@@ -772,7 +772,7 @@ static void free_pdev(struct pcifront_device *pdev)
 
 	if (pdev->gnt_ref != INVALID_GRANT_REF)
 		gnttab_end_foreign_access(pdev->gnt_ref, 0 /* r/w page */,
-					  (unsigned long)pdev->sh_info);
+					  pdev->sh_info);
 	else
 		free_page(pdev->sh_info);
